ChangeInventoriesNumericType

+++ CookManagement.VSA/Features/Inventory/GetByLowStock/LowStockInventoryResponse.cs
@@ -8,6 +8,6 @@ public sealed record LowStockInventoryResponse
     public string Product { get; init; } = String.Empty;
     public string Category { get; init; } = String.Empty;
     public string? MeasurementUnit { get; init; } = String.Empty;
-    public int CurrentStock { get; init; }
-    public int MinimumStock { get; init; }
+    public double CurrentStock { get; init; }
+    public double MinimumStock { get; init; }
 }
