+++ CookManagement.VSA/Features/Inventory/UpdateStock/InventoryItemRequest.cs
@@ -5,7 +5,7 @@ namespace CookManagement.VSA.Features.Inventory.UpdateStock
     public sealed record InventoryItemRequest
     {
         public string ProductName { get; init; } = String.Empty;
-        public int NewStock { get; init; }
+        public double NewStock { get; init; }
         public InventoryType InventoryType { get; init; }
     }
 }
